% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWorklist.R
\name{getWorklist}
\alias{getWorklist}
\title{GetWorklist}
\usage{
getWorklist(instrument = c("Agilent", "AB"), name = "worklist",
  randommethod = c("no", "position", "injection"), samplenumber = NULL,
  replication = 1, QCstep = 8, conditionQCnumber = 10,
  validationQCstep = 40, testmixstep = 40, injectionfrom = 1,
  user = "other", dir = "D:\\\\MassHunter\\\\Data\\\\SXT\\\\")
}
\arguments{
\item{instrument}{Which instrument you use?
"Agilent" or "AB", default is "Agilent".}

\item{name}{The name of worklist.}

\item{randommethod}{Which random method you want to use?
"no", "position" or "injection". Default is "no".}

\item{samplenumber}{Sample number.}

\item{replication}{Replication times.}

\item{QCstep}{QC step.}

\item{conditionQCnumber}{Condition QC number.}

\item{validationQCstep}{Validation QC step.}

\item{testmixstep}{Test mixture step.}

\item{injectionfrom}{Injection order from which? Default is 1.}

\item{user}{Default is "other".}

\item{dir}{Directory.}
}
\value{
New worklist.
}
\description{
Generate Worklist for data acquisition.
}
\details{
Generate Worklist for data acquisition.
}
\author{
Xiaotao Shen
\email{shenxt@sioc.ac.cn}
}
